requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10

# Ollama Integration
ollama>=0.1.0

# Databricks
databricks-sdk>=0.12.0
databricks-sql-connector>=2.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
pydantic>=2.0.0
tenacity>=8.2.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Type Checking
mypy>=1.5.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
